contextweaver/requirements.txt
# Core AI/ML
langchain==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.5
openai==1.10.0

# Vector Store
chromadb==0.4.22

# Document Processing
pypdf==4.0.1
python-docx==1.1.0

# Data Processing
numpy==1.26.3
pandas==2.2.0

# Ranking & Search
rank-bm25==0.2.2

# Visualization
plotly==5.18.0
networkx==3.2.1

# Web Interface
streamlit==1.30.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==8.0.0